# shared/openapi/openapi.yaml
openapi: 3.0.2
paths:
...
  /me:
    get:
      tags: &a3
        - users
      parameters: []
      operationId: users.profile
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  friendCode:
                    type: string
                  hasDivingFishImportToken:
                    type: boolean
                  hasLxnsImportToken:
                    type: boolean
                  profile:
                    nullable: true
                  idleUpdateBotFriendCode:
                    type: string
                    nullable: true
                  autoExportDivingFish:
                    type: boolean
                  autoExportLxns:
                    type: boolean
                  hasCabinetUserId:
                    type: boolean
                  autoUpdate:
                    type: boolean
                  lastScoreHash:
                    type: string
                    nullable: true
                required:
                  - id
                  - friendCode
                additionalProperties: true
    patch:
      tags: *a3
      parameters: []
      operationId: users.updateProfile
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                divingFishImportToken:
                  type: string
                  nullable: true
                lxnsImportToken:
                  type: string
                  nullable: true
                autoExportDivingFish:
                  type: boolean
                autoExportLxns:
                  type: boolean
                autoUpdate:
                  type: boolean
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  friendCode:
                    type: string
                  hasDivingFishImportToken:
                    type: boolean
                  hasLxnsImportToken:
                    type: boolean
                  profile:
                    nullable: true
                  idleUpdateBotFriendCode:
                    type: string
                    nullable: true
                  autoExportDivingFish:
                    type: boolean
                  autoExportLxns:
                    type: boolean
                  hasCabinetUserId:
                    type: boolean
                  autoUpdate:
                    type: boolean
                  lastScoreHash:
                    type: string
                    nullable: true
                required:
                  - id
                  - friendCode
                additionalProperties: true
    delete:
      tags: *a3
      parameters: []
      operationId: users.deleteMe
      responses:
        "200":
          description: "200"
  /me/prober-tokens/diving-fish:
    post:
      tags: *a3
      parameters: []
      operationId: users.getDivingFishToken
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 1
                password:
                  type: string
                  minLength: 1
              required:
                - username
                - password
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  importToken:
                    type: string
                  message:
                    type: string
                additionalProperties: true
  /me/cabinet:
    put:
      tags: *a3
      parameters: []
      operationId: users.bindCabinetQr
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                qrCode:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum:
                      - true
                required:
                  - ok
        "400":
          description: "400"
        "409":
          description: "409"
    delete:
      tags: *a3
      parameters: []
      operationId: users.unbindCabinet
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum:
                      - true
                required:
                  - ok
        "400":
          description: "400"
...
  /workers/dxnet/jobs/{jobId}/api-logs:
    post:
      tags: *a5
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      operationId: job.addApiLogs
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                logs:
                  type: array
                  items:
                    type: object
                    properties:
                      url:
                        type: string
                      method:
                        type: string
                      statusCode:
                        type: number
                      bodySize:
                        type: integer
                        minimum: 0
                        nullable: true
                    required:
                      - url
                      - method
                      - statusCode
              required:
                - logs
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
...
info:
  title: Maimai Score Hub API
  version: 1.0.0
  description: Generated from @maimai-score-hub/shared ts-rest contracts
servers:
  - url: /api/v1